Added assets for proposed homepage wireframes (.ai, .pdf, .png)
Bill brought up the idea of surfacing the full document list of the
homepage to encourage speedy discovery as opposed to linking to a
separate page. I had already sketched something out that visualized
this idea. I walked through it with Chris and Leili and got there input
and this is the resulting proposal for the homepage.